Solved code climate issues
