+++ spec/models/due_date_spec.rb
@@ -1,3 +1,5 @@
+require 'test_helper'
+
 describe "due_date_functions" do
   before(:each) do
     @deadline_type = build(:deadline_type)
@@ -131,3 +133,23 @@ describe "due_date_functions" do
     expect(DueDate.default_permission('review', 'submission_allowed')).to be == DeadlineRight::NO
   end
 end
+
+class DueDateTest < ActiveSupport::TestCase
+  describe "due_date_reminder_functions" do
+    def queue_adapter_for_test
+      ActiveJob::QueueAdapters::DelayedJobAdapter.new
+    end
+        
+    it 'enqueues remainder email in delayed job queue' do
+      expect {
+        delay.reminder()
+      }.to change(Delayed::Job.count).by(1)
+    end
+    
+    it 'enqueues remainder job in delayed job queue' do
+      expect {
+        delay.start_reminder()
+      }.to change(Delayed::Job.count).by(1)
+    end
+  end
+end
